config and minor fixes
