+++ kube-dev/1.redis.yml
@@ -1,7 +1,7 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: pcg-redis-svc
+  name: redis
   labels:
     app: pcg-test-redis
 spec:
@@ -25,12 +25,6 @@ spec:
       image: redis:5.0.4-alpine
       imagePullPolicy: Always
       command: ["redis-server"]
-      args: ["--requirepass", "$REDIS_PASSWORD"]
+      args: ["--requirepass", "dev"]
       ports:
         - containerPort: 6379
-      env:
-        - name: REDIS_PASSWORD
-          valueFrom:
-            secretKeyRef:
-              name: redis_password
-              key: dev
