+++ override/supervisord.conf
@@ -33,7 +33,7 @@ process_name=%(program_name)s-%(process_num)s
 
 ; If you want to run more than one worker instance, increase this
 ; Use 0 for debugging, start worker manually
-numprocs=0
+numprocs=1
 
 ; This is the directory from which RQ is ran. Be sure to point this to the
 ; directory where your source code is importable from
