+++ pychunkedgraph/app/config.py
@@ -18,3 +18,24 @@ class BaseConfig(object):
     # TODO what is this suppose to be by default?
     CHUNKGRAPH_TABLE_ID = "pinky100_sv16"
     # CHUNKGRAPH_TABLE_ID = "pinky100_benchmark_v92"
+
+    REDIS_URL = 'redis://localhost:6379/0'
+
+
+class DevelopmentConfig(BaseConfig):
+    """Development configuration."""
+    DEBUG = True
+
+
+class DockerDeploymentConfig(BaseConfig):
+    """Deployment configuration."""
+    REDIS_HOST = os.environ['REDIS_SERVICE_HOST']
+    REDIS_PORT = os.environ['REDIS_SERVICE_PORT']
+    REDIS_PASSWORD = os.environ['REDIS_PASSWORD']
+    REDIS_URL = f'redis://:{REDIS_PASSWORD}@{REDIS_HOST}:{REDIS_PORT}/0'
+
+
+class TestingConfig(BaseConfig):
+    """Testing configuration."""
+    TESTING = True
+    PRESERVE_CONTEXT_ON_EXCEPTION = False
